+++ finagle/buoyant/src/main/scala/com/twitter/finagle/buoyant/package.scala
@@ -1,0 +1,20 @@
+package com.twitter.finagle
+
+package object buoyant {
+
+  implicit class ParamsMaybeWith(val params: Stack.Params) extends AnyVal {
+    def maybeWith[T: Stack.Param](p: Option[T]): Stack.Params = {
+      p match {
+        case Some(t) => params + t
+        case None => params
+      }
+    }
+
+    def maybeWith(ps: Option[Stack.Params]): Stack.Params = {
+      ps match {
+        case Some(ps) => params ++ ps
+        case None => params
+      }
+    }
+  }
+}
